Bug 18436: Need to throw exception on read/modify/write or similar resolve for nonexistent property <https://bugs.webkit.org/show_bug.cgi?id=18436>
Reviewed by Maciej.

Add op_resolve_base_and_property for read/modify/write operations,
this adds a "superinstruction" to resolve the base and value of a
property simultaneously.  Just using resolveBase and resolve results
in an 5% regression in ubench, 30% in loop-empty-resolve (which is
expected).  1.3% progression in sunspider, 2.1% in ubench, with a
21% gain in loop-empty-resolve.  The only outlier is function-missing-args
which gets a 3% regression that I could never resolve.


git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@31898 268f45cc-cd09-0410-ab3c-d52691b4dbfc