Partial fix for current performance regressions.
Patch from Cameron Zwarich
Reviewed by Oliver

Fix a performance regression caused by the introduction of property
attributes to SymbolTable in r32859 by encoding the attributes and the
register index into a single field of SymbolTableEntry.

This leaves Node::optimizeVariableAccess() definitely broken, although
it was probably not entirely correct in SquirrelFish before this change.


git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@32946 268f45cc-cd09-0410-ab3c-d52691b4dbfc