Add support for for..in nodes
Reviewed by Geoff

Added two new opcodes to get_pnames and next_pname to handle iterating
over the set of properties on an object.  This iterator is explicitly
invalidated and the property name array is released on standard exit
from the loop, otherwise we rely on GC to do the clean up for us.



git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@33141 268f45cc-cd09-0410-ab3c-d52691b4dbfc