Bug 18619: Support continue, break, and return in try .. finally blocks <https://bugs.webkit.org/show_bug.cgi?id=18619>
Reviewed by Maciej

This patch replaces the current partial finally support (which uses code
duplication to achieve what it does) with a subroutine based approach.
This has a number of advantages over code duplication:
  * Reduced code size
  * Simplified exception handling as the finaliser code only exists in
    one place, so no "magic" is needed to get the correct handler for a
    finaliser.
  * When we support instruction to line number mapping we won't need to
    worry about the dramatic code movement caused by duplication

On the downside it is necessary to add two new opcodes, op_jsr and op_sret
to enter and exit the finaliser subroutines, happily SunSpider reports
a performance progression (gcc amazes me) and ubench reports a wash.

While jsr and sret provide a mechanism that allows us to enter and exit
any arbitrary finaliser we need to, it was still necessary to increase
the amount of information tracked when entering and exiting both finaliser
scopes and dynamic scopes ("with").  This means "scopeDepth" is now
the combination of "finaliserDepth" and "dynamicScopeDepth".  We also
now use a scopeContextStack to ensure that we pop scopes and execute
finalisers in the correct order.  This increases the cost of "with" nodes
during codegen, but it should not be significant enough to effect real
world performance and greatly simplifies codegen for return, break and
continue when interacting with finalisers.



git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@33227 268f45cc-cd09-0410-ab3c-d52691b4dbfc