Bug 18642: Iterator context may get placed into the return register, leading to much badness <https://bugs.webkit.org/show_bug.cgi?id=18642>
Reviewed by Maciej

To prevent incorrectly reusing what will become the result register for
eval and global code execution, we need to request and ref the destination
in advance of codegen.  Unfortunately this may lead to unnecessary copying,
although in future we can probably limit this.  Curiously SunSpider shows
a progression in a number of tests, although it comes out as a wash overall.

This also fixes one of the regressions in run-javascriptcore-tests.



git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@33232 268f45cc-cd09-0410-ab3c-d52691b4dbfc