Bug 19024: SQUIRRELFISH: ASSERTION FAILED: activation->isActivationObject() in Machine::unwindCallFrame <https://bugs.webkit.org/show_bug.cgi?id=19024>
Reviewed by Maciej

This fixes a number of issues.  The most important is that we now check every register
file for tainting rather than just looking for function register files as that was
insufficient. Additionally guarded against implicit re-entry into Eval code.

Also added a few additional assertions to reduce the amout of time between something
going wrong and us seeing the error.


git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@33438 268f45cc-cd09-0410-ab3c-d52691b4dbfc