Bug 19076: SquirrelFish: RegisterFile can be corrupted if implictly reenter global scope with no declared vars <https://bugs.webkit.org/show_bug.cgi?id=19076>
Reviewed by Geoff

Don't delay allocation of initial global RegisterFile, as we can't guarantee we will be able
to allocate the global 'this' register safely at any point after initialisation of the Global
Object.

Unfortunately this initial allocation caused a regression of 0.2-0.3%, however this patch adds
support for the static slot optimisation for the global Math object which brings it to a 0.3%
progression.



git-svn-id: http://svn.webkit.org/repository/webkit/branches/squirrelfish@33914 268f45cc-cd09-0410-ab3c-d52691b4dbfc